% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core.R
\name{DLFM_Dunson_fixed}
\alias{DLFM_Dunson_fixed}
\title{DLFM_Dunson_fixed}
\usage{
DLFM_Dunson_fixed(Y, X, RE = c("additive", "multiplicative"), R = 2,
  dist = "Exponential", gammapriors = c(2, 1), avail = matrix(1,
  dim(Y)[1], dim(Y)[2]), burn = 1000, nscan = 5000, odens = 100, kappas)
}
\arguments{
\item{Y}{relational array of relations}

\item{X}{Time x n x n x p covariate array}

\item{RE}{random effect to be included ("additive" and/or "multiplicative)}

\item{R}{dimension of the multiplicative effects}

\item{dist}{standard Exponential or squared Exponential}

\item{gammapriors}{inverse-gamma shape and scale parameters for (s2, beta, theta, d)}

\item{avail}{= Time x n matrix reperesenting the availibility of nodes (1 if avail, 0 if structural NA)}

\item{burn}{burn in for the Markov chain}

\item{nscan}{number of iterations of the Markov chain (beyond burn-in)}

\item{odens}{output density for the Markov chain}

\item{kappas}{P+2 length vector of GP length parameters}
}
\value{
Final estimate of the parameters
}
\description{
MCMC algorithm using Gibbs sampling for each variable with structural NA
}
