% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core.R
\name{rbeta_fc}
\alias{rbeta_fc}
\title{rbeta_fc}
\usage{
rbeta_fc(X, beta, theta, UDU, Y, cinv, tau_p, s2)
}
\arguments{
\item{X}{Time x n x n x p covariate array}

\item{beta}{a p by 1 vector of coefficients}

\item{theta}{a T x n matrix of nodal effects}

\item{UDU}{Time-length list of UDU}

\item{Y}{relational array of relations}

\item{cinv}{inverse of Gaussian process covariance matrix}

\item{tau_p}{scale parameter of Gaussian process covariance}

\item{s2}{variance of normal error}
}
\value{
One sample estimate of beta from normal distribution
}
\description{
Sample beta
}
