% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core.R
\name{rs2_fc2}
\alias{rs2_fc2}
\title{rs2_fc2}
\usage{
rs2_fc2(XB, theta, UDU, Y, a, b)
}
\arguments{
\item{XB}{Time-length list of XB matrix}

\item{theta}{a T x n matrix of nodal effects}

\item{UDU}{Time-length list of UDU}

\item{Y}{relational array of relations}

\item{a}{inverse-gamma shape parameter}

\item{b}{inverse-gamma scale parameter}
}
\value{
One sample estimate of sigma^2 from inverse-gamma distribution
}
\description{
Sample the variance in Normal error
}
