% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core.R
\name{rtauu_fc}
\alias{rtauu_fc}
\title{rtauu_fc}
\usage{
rtauu_fc(U, a, b)
}
\arguments{
\item{U}{current value of d}

\item{a}{inverse-gamma shape parameter}

\item{b}{inverse-gamma scale parameter}
}
\value{
One sample estimate of tau_r from inverse-gamma distribution
}
\description{
Sample the variance in d
}
